+++ packages/ui/src/provider/app-wrapper.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { Navbar } from "../components/navbar";
 import { Providers } from "./next-theme";
+import { Toaster } from "../components/sonner";
 
 interface Props {
   children: React.ReactNode;
@@ -11,6 +12,7 @@ export function AppWrapper({ children }: Props) {
     <Providers>
       <Navbar />
       {children}
+      <Toaster />
     </Providers>
   );
 }