+++ apps/ios-mockups/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ios-mockups
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/ios-mockups@0.0.4
+
 ## 0.0.2
 
 ### Patch Changes