+++ packages/ios-mockups/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zuude-ui/ios-mockups
 
+## 0.0.4
+
+### Patch Changes
+
+- Scrollable feature on iphone mockup
+
 ## 0.0.3
 
 ### Patch Changes